requirements.txt
#requirements
streamlit >= 1.45.0 
pillow >= 11.2.1
